Autenticação por senha criado.

+++ portfolio/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render,redirect
 from .models import Tecnologia,Competencia,Formacao,Licenciatura,Projeto,Voluntariado
 from .forms import CompetenciaForm, TecnologiaForm, FormacaoForm, ProjetoForm, VoluntariadoForm
+from django.contrib.auth.decorators import login_required
 
+### VIEWS Principais ============================================================================================
 
 def main_view(request):
     return render(request, 'portfolio/index.html')
@@ -9,9 +11,48 @@ def main_view(request):
 
 def tecnologias_view(request):
     tecnologias = Tecnologia.objects.all()
-    return render(request, 'portfolio/tecnologias.html',{'tecnologias':tecnologias})
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request, 'portfolio/tecnologias.html',{'tecnologias':tecnologias, 'gestor': gestor,})
 
+def competencias_view(request):
+    competencias = Competencia.objects.all()
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request, 'portfolio/competencias.html', {
+        'competencias': competencias,
+        'gestor': gestor,
+    })
+
+def formacoes_view(request):
+    formacoes = Formacao.objects.all()
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request,'portfolio/formacoes.html',{'formacoes':formacoes,'gestor': gestor,} )
+
+def licenciatura_view(request):
+    licenciaturas = Licenciatura.objects.all()
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request,'portfolio/licenciatura.html',{'licenciaturas':licenciaturas,'gestor': gestor,} )
+
+def projetos_view(request):
+    projetos = Projeto.objects.all()
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request, 'portfolio/projetos.html', {
+        'projetos': projetos,
+        'gestor': gestor,
+    })
+
+def voluntariado_view(request):
+    voluntariados = Voluntariado.objects.all()
+    gestor = request.user.groups.filter(name='gestor-portfolio').exists()
+    return render(request,'portfolio/voluntariado.html',{'voluntariados':voluntariados,'gestor': gestor,} )
+
+
+
+### VIEWS PARA Criar Objetos ============================================================================================
+
+@login_required
 def tecnologia_nova_view(request):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     if request.method == 'POST':
         form = TecnologiaForm(request.POST)
         if form.is_valid():
@@ -25,13 +66,12 @@ def tecnologia_nova_view(request):
     })
 
 
-def competencias_view(request):
-    competencias = Competencia.objects.all()
-    return render(request, 'portfolio/competencias.html', {
-        'competencias': competencias,
-    })
 
+
+@login_required
 def competencia_nova_view(request):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     if request.method == 'POST':
         form = CompetenciaForm(request.POST)
         if form.is_valid():
@@ -45,11 +85,10 @@ def competencia_nova_view(request):
     })
 
 
-def formacoes_view(request):
-    formacoes = Formacao.objects.all()
-    return render(request,'portfolio/formacoes.html',{'formacoes':formacoes} )
-
+@login_required
 def formacoes_nova_view(request):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     if request.method == 'POST':
         form = FormacaoForm(request.POST)
         if form.is_valid():
@@ -62,15 +101,11 @@ def formacoes_nova_view(request):
         'form': form,
     })
 
-def licenciatura_view(request):
-    licenciaturas = Licenciatura.objects.all()
-    return render(request,'portfolio/licenciatura.html',{'licenciaturas':licenciaturas} )
 
-def projetos_view(request):
-    projetos = Projeto.objects.all()
-    return render(request,'portfolio/projetos.html',{'projetos':projetos} )
-
+@login_required
 def projetos_nova_view(request):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     if request.method == 'POST':
         form = ProjetoForm(request.POST)
         if form.is_valid():
@@ -83,11 +118,12 @@ def projetos_nova_view(request):
         'form': form,
     })
 
-def voluntariado_view(request):
-    voluntariados = Voluntariado.objects.all()
-    return render(request,'portfolio/voluntariado.html',{'voluntariados':voluntariados} )
 
+
+@login_required
 def voluntariado_nova_view(request):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     if request.method == 'POST':
         form = VoluntariadoForm(request.POST)
         if form.is_valid():
@@ -101,38 +137,56 @@ def voluntariado_nova_view(request):
     })
 
 
-### VIEWS PARA O DELETE ----------------------------------
+### VIEWS PARA O DELETE ============================================================================================
+@login_required
 def tecnologia_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     tecnologia = Tecnologia.objects.get(id = id)
     if request.method == 'POST':
         tecnologia.delete()
     return redirect('tecnologias')
 
+@login_required
 def competencia_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     competencia = Competencia.objects.get(id = id)
     if request.method == 'POST':
         competencia.delete()
     return redirect('competencias')
 
+@login_required
 def formacao_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     formacao = Formacao.objects.get(id = id)
     if request.method == 'POST':
         formacao.delete()
     return redirect('formacoes')
 
+@login_required
 def licenciatura_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     licenciatura = Licenciatura.objects.get(id = id)
     if request.method == 'POST':
         licenciatura.delete()
     return redirect('licenciatura')
 
+@login_required
 def projeto_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     projeto = Projeto.objects.get(id = id)
     if request.method == 'POST':
         projeto.delete()
     return redirect('projetos')
 
+@login_required
 def voluntariado_eliminar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     voluntariado = Voluntariado.objects.get(id = id)
     if request.method == 'POST':
         voluntariado.delete()
@@ -140,8 +194,11 @@ def voluntariado_eliminar_view(request, id):
 
 
 
-### VIEWS DE EDIÇÃO -------------------------------------------
+### VIEWS PARA Edição ============================================================================================
+@login_required
 def tecnologia_editar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     tecnologia = Tecnologia.objects.get(id = id)
     if request.method == 'POST':
         form = TecnologiaForm(request.POST, request.FILES, instance=tecnologia)
@@ -156,8 +213,10 @@ def tecnologia_editar_view(request, id):
         'tecnologia': tecnologia,
     })
 
-
+@login_required
 def competencia_editar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     competencia = Competencia.objects.get(id = id)
     if request.method == 'POST':
         form = CompetenciaForm(request.POST, instance=competencia)
@@ -172,8 +231,10 @@ def competencia_editar_view(request, id):
         'competencia': competencia,
     })
 
-
+@login_required
 def formacao_editar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     formacao = Formacao.objects.get(id = id)
     if request.method == 'POST':
         form = FormacaoForm(request.POST, instance=formacao)
@@ -188,8 +249,10 @@ def formacao_editar_view(request, id):
         'formacao': formacao,
     })
 
-
+@login_required
 def projeto_editar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     projeto = Projeto.objects.get(id = id)
     if request.method == 'POST':
         form = ProjetoForm(request.POST, request.FILES, instance=projeto)
@@ -204,8 +267,10 @@ def projeto_editar_view(request, id):
         'projeto': projeto,
     })
 
-
+@login_required
 def voluntariado_editar_view(request, id):
+    if not request.user.groups.filter(name='gestor-portfolio').exists():
+        return redirect('landingPage')
     voluntariado = Voluntariado.objects.get(id = id)
     if request.method == 'POST':
         form = VoluntariadoForm(request.POST, request.FILES, instance=voluntariado)
